refactor(documents): extract saveDocument into shared service
Part of JARVIS-698
